client/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
config~=0.5.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic_settings==2.6.0
azure-storage-queue==12.12.0
